docs: Add README for behavior test and ops benchmarks
