+++ .env.example
@@ -1,0 +1,10 @@
+# fs
+OPENDAL_FS_TEST=false
+OPENDAL_FS_ROOT=/path/to/dir
+# s3
+OPENDAL_S3_TEST=false
+OPENDAL_S3_BUCKET=<bucket>
+OPENDAL_S3_REGION=<region>
+OPENDAL_S3_ENDPOINT=<endpoint>
+OPENDAL_S3_ACCESS_KEY_ID=<access_key_id>
+OPENDAL_S3_SECRET_ACCESS_KEY=<secret_access_key>
