+++ Cargo.toml
@@ -9,6 +9,9 @@ repository = "https://github.com/datafuselabs/opendal"
 keywords = ["storage", "data", "s3"]
 categories = ["filesystem"]
 
+[lib]
+bench = false
+
 [[bench]]
 name = "ops"
 harness = false
@@ -34,5 +37,6 @@ uuid = { version = "0.8.2", features = ["serde", "v4"] }
 anyhow = "1.0"
 rand = "0.8.5"
 sha2 = "0.10.1"
-criterion = { version = "0.3", features = ["async", "async_tokio"] }
+criterion = { version = "0.3", features = ["async", "async_tokio", "html_reports"] }
 tokio = { version = "1.16.1", features = ["full"] }
+dotenv = "0.15.0"
